[Postgresql] Update Puppet to 9.5
The migration has been completed, just need puppet not to destroy my hard
work.  All references to 9.3 have been updated to 9.5.  I've removed the
control package explicit install as it's not parameterized and unneeded as
it's a a recommended package by postgresql-X.Y and will be pulled in auto-
matically.  Dc_postgres is a bit smelly so I will tidy it up next.