[RAID] Fix Up Broken Code
Came to my attention when looking at Puppet 4... So apparently the madam module does stuff when
the software_raid fact is defined, which amazingly it always is as it will always default to an
empty string.  This makes the code simpler, only sets the fact when there is something to
report and also, for bonus points, returns an array for structured fact goodness.  Backwards
compatible insofar as the actual value was never actually used.