[Puppet4] Upgrade Galera (#2939)
The new P4 friendly MySQL module is incompatible with our fork of Galera (note to
self - push modifications upstream).  This rebases Nick's changes on the upstream
master, and also adds into hiera variables that now need to be specified.  These
should be unchanged from the previous revision.  Tested in vagrant, but obviously
given the critical nature will be no-oped first.