[Icinga2] Add Telegraf Service Check (#2947)
Aids in tracking where upgrades need to be performed etc which prevent the service
from running.  Just a simple check to warn if more than one process is running and
needs some non-urgent attention or go critical if none are running and we cannot
see what the box is doing.