[docs] Switch to nginx with working vhosts: (#2949)
This commit updates the nginx submodule to the voxpupuli repo
(https://github.com/jfryman/puppet-nginx.git is just a
reference to the voxpupuli repo, so nothing should change
here) and checks out to a commit before the one that reverted
the hostfix from
https://github.com/voxpupuli/puppet-nginx/pull/624

This fixes the problem with nginx trying to create a vhost
configuration before installing the nginx package. This is
caused by concat not requiring the underlying nginx file
hierarchy.

This solution needs to be reverted when concat devs come up
with a working implementation of the autorequire function
in `lib/puppet/type/concat_file.rb`. The current solution
from ba643617a2a03f0588a0ec6625824a707528d9f0 (L112-L115)
is not working and the original approach:
https://github.com/puppetlabs/puppetlabs-concat/pull/319
has been abandoned at some point.

For more details see the ticket:
https://datacentred.atlassian.net/browse/PD-2567.