[Ceph] Update Module To 1.5.3 (#2977)
Although technically this version only officially supports Jewel it is
also able to manage Hammer through a tweaking of the parameters.  This
gets us to where we need to be in a few weeks (with the exception of
the libjemalloc LD_PRELOAD) as well as adding in functionality to actually
provision the QCT gear.  Adds in the correct SKU configuration.