[RabbitMQ] Fix Configuration (#2986)
Ensures the file rlimit is the same as it was previously.  The default environment
variables have been updated from RABBITMQ_NODE_PORT to NODE_PORT, as I haven't got
the slightest clue how rabbitmq works I figure it is safest to add in the variable
that was removed.  Also the SSL paths have been updated.