[MongoDB] Fix Connection Validation (#2990)
The updated module tries to connect to the mongo server to validate it's alive.  Sadly
we hard code the bind IP to '' which causes an exception when passed to IPAddr.new.  This
sets the bind IP instead to be 0.0.0.0 (listen to all the things) then the puppet code
takes care of turning this into 127.0.0.1 to do the validation.