[TFTP] Back Upstream (#3076)
Much like the last N of these remove our hacks and put them in wrapper classes so
that we can point at the maintained code repos.  Cleaned out a load of unused hiera
and classes.  Added ruby-rubyipmi dependency to the nameservers, the pervious code
got orphaned at some point and the gem being installed is the same as available via
apt _sigh_