[Galera] Re-enable Status Check (#3082)
As Nick so kindly points out this is used by haproxy to decide whether to punt
requests at the server.  I'm re-enabling this functionality, however it will
be rolled out gently-gently; manually adding the statuscheck user as a system
account on all osdbmq nodes, provisioning one and seeing what happens, so we
at least have 2 if things go wrong.