[Icinga2] Disable Connection Tracking (#3125)
Observe the following:

    simon@osnet0:~$ time sudo ./check_conntrack_2
    OK: flows=366976 max_flows=1048576 | flows=366976;838860;943718;0;1048576

    real0m44.863s
    user0m0.692s
    sys0m11.996s

We are missing 30 seconds, which could be kernel contention over locks, in other words we
are impacted by normal operation, and normal operation is impacted by us.  Leave connection
tracking in a state of quantum superposition by not observing it in the hope the cat will
live through January.