[Keystone] Disable classes which attempt to manage Keystone resources (#3179)
The whole process around creating resources in the Keystone catalogue
for various things (services, endpoints, users) needs refactoring, and
right now it's in a slightly broken state that means we can't provision
a new control node cleanly until this has been re-thought.

As a workaround, this commit disables classes which attempt to manage
Keystone resources for various services.  It won't have any effect on
service as these are all already registered.