[Nova] Configure nova-scheduler appropriately for Active/Active (#3175)
The recommended setting and its rational is best described in this
comment here:

https://bugzilla.redhat.com/show_bug.cgi?id=1104219

Specifically:

``` When using Active/Active, the following configuration option needs
to be set in nova.conf: scheduler_host_subset_size.  This option
defaults to 1.  I would recommend setting it to a default of "10" when
using Active/Active.  However, it's something that should be allowed
to be tweaked if needed.
```

This commit makes the recommended change.