[Staging] Puppet Agent on DNS Master (#3186)
Creates a system tftp user on which the directory creation relies.  We cannot
do clever things here as the ::tftp class is monolithic.  Adds in DNS zones to
mirror those added via DNS bootstrap.  DHCP configuration is faked (albeit
accurately enough for testing) as staging in essence needs to be completely
statically addressed.  A quick bug fix in the puppet 4 configuration code which
wasn't actually matching the domain correctly for example.com.