[Gateway] HAP and Keepalived (#3201)
Add in configuration for puppet, puppetdb and foreman services in staging.  I'll
add the rest in when there are addresses allocated for the other services. Move
the keepalived configuration from manifest to hiera as it's domain specific and
needs to be twiddled between environments.  HAP was pinned at some archaic version
not present in Xenial so unpin it, apt-cache policy appears to suggest it will
always install >= 1.5 anyway.  Don't merge willy nilly as this has the ability
to kill off all our monitoring and logging :)