[Puppet] Fix Circular Deps on Production (#3212)
Adding directory creation in the setup stage created a dependency on the root
user resource in the main stage and kaboom.  Interestingly when I looked at
the preinst script for the puppetserver package it basically runs a usermod
on an existing puppet user to hijack it and correct the home directory.  Using
this magical hidden information if we make deployment key installation dependant
on server installation things 'should just work' (tm).