[Icinga] Update Anti Affinity (#3222)
The check ceased working with keystone v3 so this fixes it up again.  I've updated
the interface to be more in sync with other new(er) checks e.g. project rather than
tenant.  Also deletes the python2.7 horror code so this will work on Xenial once the
upgrade push starts.