[Icinga2] Hack Around Alan's Snowflake Part 3 (#3237)
Final piece of the pie.  Move from a single check to a check applied to each
mounted filesystem.  At this point we are able to reject applying the service
check to specific mounts on specific hosts.  As an added bonus I believe that
tracefs will now be ignored as that's not whitelisted in the metadata generator.
I've renamed the check to be file system rather than disk as it's more accurate.