Add the new AggregateImagePropertiesIsolation filter: (#3276)
* Add the new AggregateImagePropertiesIsolation filter:

A hacky way of adding the new implementation of the
`AggregateImagePropertiesIsolation` scheduler filter.
Should be good for the time we are on Kilo. And then if the
Mitaka version of the filter proves to be broken as well,
this can be ported pretty easily.
I'm submitting this change mostly to get some more eyes to see
and review the filter's code, so any criticism is welcome. ;)
Also I want to probe if you thing this code is worth submitting
upstream over the existing implementation which seems to have
a few problems even with the latest changes on the master
branch.

Since this change can potentially break the ability to
scheduler instances, DO NOT MERGE!

Coauthored with Peter Tar.

* Fix file mode not being a string.

* Fix pylint issues:

* Disable import of modules.
* Deal with the "Method could be a function (no-self-use)"
warning.
* Disable "too-few-public-methods" check.

* Move the filter's code to `dc_openstack`.