[Monitoring] Modifying iptables check script (#3337)
* [Monitoring] Modifying iptables check script

Since the keystoneauth1 module is missing from osnet nodes it's not
possible to use the openstack api. The script has been modified to use
ip netns commands instead. It can't tell if the router is active or
standby but otherwise it works. It is a temporary solution until
Openstack gets upgraded.

* Update dc_icinga2.yaml