[Icinga2] Smarter OpenStack Check (#3352)
My naïve implementation on pgrep just returned the first match which was
the numerically lowest PID.  This version attempts to return the group
leader e.g. the first PID matching the process whose parent is not also
in the list.  This way the child PID check is more likely to succeed.