[Icinga2] Smart Proxy Check Script (#3356)
Push out to the DNS masters before enabling the check.  Ensures interfaces in
Foreman are correctly configured in the DNS, DHCP and PXE services.  Replaces
the old Jarvis classic which was hard coded to python 2.7... Oh it also does
the unthinkable and makes this a Nagios plugin and not a cron job