[DHCP] Rejig Ranges (#3380)
We are at 96% utilisation on storage public.  This changes the dynamic range to be a more
sane /28 rather than an arbitrary pretty number.  Same goes for platform services.  BMC
and infrastructure networks are properly run so don't need separate ranges, these have been
expanded to exclude the highest and lowest /28 prefixes.