[NTP] Fix time on external hosts (#3407)
Interfaces on the 10.0.0.0/8 are specified as global options, which breaks any other kit not in core or sal01. This change moves the listen config stanzas to the domain specifics, removes the listen on any other kit (we have no other ntp servers). Also adds extra external pool targets to the global configs to 4 from 2 and updates the NTP module (tested works)