[Security] Additional Unattended Upgrade nodes (#3420)
This change adds the unattended profile to another set of nodes. Nodes have been chosen that are part of a cluster for now, or public facing and recently updated (so minimal deltas). Happy with the functionality so far, so increasing scope. Probably the last now before globally enabling (bar HV's)