[Website] Add SSL to www.datacentred.co.uk (#3422)
This change adds SSL via letsencrypt for www.datacentred.co.uk. I have a feeling we might hit an issue with the way the IP has been rotuted to the box, but we'll give it a go :)
Doesn't change any underlying nginx config (although it'll need to bind to port 80 and 443 during cert request). Hihly likely some tuning will be needed but difficult to test it fully in vagrant due to .test being an invalid public TLD. Also refactors some of the wordpress puppet layout to make it more generic

Addresses PD-3050