[VPN] Downgrade for 2.3 support and add DCIM (#3438)
This change removes some configuration that was causing issues with 2.3 clients, in the sake of getting a workable replacement. Once all VPN endpoints are under puppet curation, we can then increase back to 2.4 level security. We still retain the benefits of the 2.4 audited codebase this way but ensure compatibility for users that need to access Opengear based VPNs