[Telegraf] Fixing the telegraf docker plugin (#3459)
Currently the settings do not get applied, so they are being moved to
the role yamls (oscontrol,jenkins).

Also, in order for the plugin to work the user telegraf needs to be a
member of the docker group. The telegraf profile is modified to allow
the user to have different memberships based on the groups available on
the node.