[Ceph] Configuration Updates (#3476)
So at some point in the distant past some pillock in the Swift team decided to
scope the catalog URLs for public and internal endpoints.  The API version wasn't
bumped, as as far as I can guess the scoping is prefixed by the catchy AUTH_.
Now all the tooling (especially tempest) moved on to assume the new layout.  Ceph
however are still pedalling the archaic version (probably a money grab by RedHat).
Luckily due to open source we could see what special hacks they charge for and
activate them manually.  This is already live and working, just needs for the
configuration not to get inadvertantly blown away.