[Icinga2] Export Block Device Names (#3519)
Part 1 of making Nick a happy bunny.  This exports the block device model with
the path which we can then use in icinga2 to build the per disk check description.
Thus we can at a glance see whether something is a journal or not given device
enumeration can happen in any random order.