[ElasticSearch] Set MAX_LOCKED_MEMORY to 'unlimited'
ES does a memory lock check on startup and under some scenarios this can
fail if it's not set to 'unlimited'.

Given we're reserving 50% of memory for usage by ES, and it's deployed
on dedicated boxes with 32GB, if it can't lock all the memory it's
requesting then we've another problem that needs dealing with anyway.