[Elasticsearch] Add Monit to ensure uptime of service
Due to recent issues with elasticsearch falling over, I'm taking a belt and braces approach. This change adds the monit upstream module, a profile and a monit check to make sure elasticsearch is always running. We'll get monitoring if it does fall over, but at least it'll kick it back to life.