[Icinga2] NTP Improvements Part I (#3596)
So previously we made an assumption that all servers would reside in a domain
and have a local NTP server.  That was until we started using VMs... and then
monitorable machines in the ether e.g. transit monitoring and VPNs.  Rather
than selectively ignore things we should be able to monitor clock drift on
everything right?  Indeed, the easiest thing I can think of is to parse the
ntp configuration, export a fact and add to the host metadata, thus making the
service check dynamic.  This caters for the fact and updating host definitions.