[Nginx] Update Cert on WWW (#3601)
Our friends in the government have correctly noticed that the apex for datacentred.io
is CNAMED (and squished into an A record) to datacentred.co.uk.  First thing that
happens is your browser barfs that the certificate SAN is wrong.  This change adds in
SANs for all of the domains that point at the server.

I could add in a vhost raising a 301 for all the other domains, but we'd still need
the multi-domain SSL cert anyway if user X tried to connect to https://datacentred.org
directly.