[Icinga2] Refresh Postgres Check
Was still seeing some flapping even after upping the thresholds, transpires
that it wasn't the receipt of xlogs but the application of them that was
causing the warnings.  These are now bound to the same thresholds as the
receipt.  Did some de-duplication cleanup and other general spring cleaning
to make the check prettier.  Also added in some performance data so we can
actual quantify what we should set the thresholds to based on historical
time-series data.