[Elasticsearch] Backup Script Upgrade (#3629)
Fixes the formatting of the snapshot name, just to keep things uniform.  Add
in some logging so we know what's being kept, what's being deleted and what
state the snapshots are in.  As this goes to stderr it will end up in the cron
email sent out.  Finally we need to establish a new HTTP connection with each
operation as elastic search doesn't keep it open across multiple deletes.