[Haproxy] Capture real client ip addresses (#3663)
Just as the title says, this log format captures the real client ip
addresses and in general produces a nice output. It's the work of
Mariusz.
Tested in vagrant with keystone, glance and nova. It should work with
the rest as well.
The grok filters are coming in the next pr.

PD-3278