[Monitoring] Hide ssl warning of heat check (#3656)
The heat check produces an ssl warning upon accessing the openstack api.
```InsecurePlatformWarning: A true SSLContext object is not available.
This prevents urllib3 from configuring SSL appropriately and may cause
certain SSL connections to fail. You can upgrade to a newer version of
Python to solve this.```
It is not solved by installing 'requests[security]'. Since the check is
working and these are only warning messages, I decided to just suppress
them..